Add route change and page unload protection for exam creation form
- Implemented route change prevention when form is unsaved
- Added confirmation dialog for leaving unsaved exam creation form
- Removed redundant beforeunload event listeners from Step1 and Step2 components
- Centralized form state protection logic in the main CreateExam page component
