+++ src/components/create-exam/step1.tsx
@@ -267,7 +267,7 @@ interface Step1Props {
 export const Step1 = ({ onNext }: Step1Props) => {
   const {
     control,
-    formState: { errors, isDirty, isSubmitted },
+    formState: { errors },
     trigger,
     watch,
   } = useStep1Form();
@@ -305,17 +305,6 @@ export const Step1 = ({ onNext }: Step1Props) => {
     setActiveQuestionIndex((prev) => Math.min(prev > 0 ? prev - 1 : prev, fields.length - 2));
   };
 
-  // “beforeunload” => form kaydedilmediyse uyar
-  useEffect(() => {
-    const handleBeforeUnload = (e: BeforeUnloadEvent) => {
-      if (isDirty && !isSubmitted) {
-        e.preventDefault();
-        e.returnValue = "";
-      }
-    };
-    window.addEventListener("beforeunload", handleBeforeUnload);
-    return () => window.removeEventListener("beforeunload", handleBeforeUnload);
-  }, [isDirty, isSubmitted]);
 
   useEffect(() => {
     if (questions.length > 0 && questionRefs.current && questionRefs.current[activeQuestionIndex]) {
